requirements.txt
asgiref==3.3.4
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
cymem==2.0.5
Django==3.2.3
django-environ==0.4.5
djangorestframework==3.12.4
django-filter==2.1.0
emoji==1.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
gunicorn==20.1.0
idna==2.10
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
murmurhash==1.0.5
nltk==3.6.2
numpy==1.20.3
packaging==20.9
pathy==0.5.2
Pillow==8.2.0
preshed==3.0.5
psycopg2-binary==2.8.6
pydantic==1.7.4
pyparsing==2.4.7
pytz==2021.1
regex==2021.4.4
requests==2.25.1
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.6
sqlparse==0.4.1
srsly==2.4.1
textblob==0.15.3
thinc==8.0.4
tqdm==4.61.1
typer==0.3.2
urllib3==1.26.5
wasabi==0.8.2
